+++ ChangeLog.md
@@ -1,5 +1,31 @@
 ## change log ##
 
+* 2016-04-18: 0.3.1:
+  - upload file: use the following properties for possible ontology names:
+    RDFS.label, Omv.name, DCTerms.title, DC_11.title, DC_10.title
+  - The reported info upon uploading a file looks thus:
+    ```
+	{
+		"filename": "1461043897684.rdf",
+		"format": "rdf",
+		"possibleOntologyUris": {
+			"http://example.org/ont1": {
+				"explanations": [
+					"Namespace associated with empty prefix but with no trailing separators",
+					"Value of xml:base attribute"
+				],
+				"names": [
+					{
+						"propertyUri": "http://omv.ontoware.org/2005/05/ontology#name",
+						"propertyValue": "Test ontology"
+					}
+				]
+			}
+		},
+		"userName": "carueda"
+	}
+    ```
+  
 * 2016-04-16: 0.3.1:
   - also handle previously uploaded file in PUT /ont